requirements.txt
astroid==2.8.4
attrs==21.2.0
aws-cdk.assets==1.130.0
aws-cdk.aws-applicationautoscaling==1.130.0
aws-cdk.aws-autoscaling-common==1.130.0
aws-cdk.aws-cloudformation==1.130.0
aws-cdk.aws-cloudwatch==1.130.0
aws-cdk.aws-codeguruprofiler==1.130.0
aws-cdk.aws-codestarnotifications==1.130.0
aws-cdk.aws-ec2==1.130.0
aws-cdk.aws-ecr==1.130.0
aws-cdk.aws-ecr-assets==1.130.0
aws-cdk.aws-efs==1.130.0
aws-cdk.aws-events==1.130.0
aws-cdk.aws-iam==1.130.0
aws-cdk.aws-kms==1.130.0
aws-cdk.aws-lambda==1.130.0
aws-cdk.aws-logs==1.130.0
aws-cdk.aws-rds==1.130.0
aws-cdk.aws-redshift==1.130.0
aws-cdk.aws-s3==1.130.0
aws-cdk.aws-s3-assets==1.130.0
aws-cdk.aws-sam==1.130.0
aws-cdk.aws-secretsmanager==1.130.0
aws-cdk.aws-signer==1.130.0
aws-cdk.aws-sns==1.130.0
aws-cdk.aws-sqs==1.130.0
aws-cdk.aws-ssm==1.130.0
aws-cdk.cloud-assembly-schema==1.130.0
aws-cdk.core==1.130.0
aws-cdk.custom-resources==1.130.0
aws-cdk.cx-api==1.130.0
aws-cdk.region-info==1.130.0
cattrs==1.8.0
constructs==3.3.161
isort==5.9.3
jsii==1.41.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
platformdirs==2.4.0
publication==0.0.3
pylint==2.11.1
python-dateutil==2.8.2
six==1.16.0
toml==0.10.2
typing-extensions==3.10.0.2
wrapt==1.13.3
